requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
clip==1.0
colorama==0.4.6
filelock==3.8.0
ftfy==6.1.1
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.0.0
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.10.0
